Fix traverse restarts.
When the leaf is locked and the target key is not found (insert) or
found (delete), traverse has to lock, check the path and restart if necessary.

Existing code didn't work because traverse result (which is -EEXIST or -ENOENT
in out case), was used for two incompatible purposes:

    - indicate success or failure or traverse,
    - indicate success or failure of operation (insertion or deletion).

Introduce path->rc for the latter.

Signed-off-by: Nikita Danilov <danilov@gmail.com>