Delete: support variably-sized keys.
When nodes are merged during deletion, a separating key
higher in the tree needs to be updated (SEPCHG,
delete_update()). As the new separating key can be longer
than the old one, deletion might result in a split and
a new root.

For now, only merge nodes that share the immediate parent.

Signed-off-by: Nikita Danilov <danilov@gmail.com>