leaf_search(): fix a race.
leaf_search() updates the key and the value in the path record. If the path is
not valid after the search, the traversal is retried, but with the
wrong (updated) key.

Fix this by providing a scratch key buffer in t2_rec. Copy the key to the
scratch buffer on first traversal and copy it back on each retry.

Signed-off-by: Nikita Danilov <danilov@gmail.com>