WAL: fix out-of-order log buffer write completion.
Log buffers can be written out of (lsn-)-order. This means
that ->max_synced saved in buffer headers are not necessary
monotone and this breaks wal_log_sync(), written at
the time when buffers were written strictly in order.

Fix this by remembering ->max_synced (and ->start) in
an additional array en->persisted[].

Signed-off-by: Nikita Danilov <danilov@gmail.com>